+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file. See [versionize](https://github.com/versionize/versionize) for commit guidelines.
 
+<a name="0.1.0"></a>
+## [0.1.0](https://www.github.com/ggwozdz90/translation-api-client/releases/tag/v0.1.0) (2025-03-25)
+
+### Features
+
+* add DTOs for translation requests and responses ([54bd880](https://www.github.com/ggwozdz90/translation-api-client/commit/54bd880cd1db61b3af029f4f490ac5184617f1bd))
+* implement translation API client ([a44613e](https://www.github.com/ggwozdz90/translation-api-client/commit/a44613ef9fb69983fdab508cb369642e0114901c))
+
 <a name="0.0.1"></a>
 ## [0.0.1](https://www.github.com/ggwozdz90/csharp-project-template/releases/tag/v0.0.1) (2025-02-03)